wire double precision clocks through
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@5311 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/sound/ym2151_opm.c
@@ -2013,7 +2013,7 @@ void OPM_SetIC(opm_t *chip, uint8_t ic)
     }
 }
 
-void OPM_Reset(opm_t *chip, uint32_t clock)
+void OPM_Reset(opm_t *chip, double clock)
 {
 #ifdef PINMAME
     void(*irqhandler)(int irq) = chip->irqhandler;
@@ -2028,7 +2028,7 @@ void OPM_Reset(opm_t *chip, uint32_t clock)
     }
     OPM_SetIC(chip, 0);
 #ifdef PINMAME
-    if (clock != 0) // init?
+    if (clock != 0.) // init?
     {
         chip->vgm_idx = vgm_open(VGMC_YM2151, clock);
     }
